pyproject.toml
[tool.poetry]
name = "llm-codellama"
version = "0.0.9000"
description = ""
authors = ["Michihiro Nakamura <nakamichi@nakamichi.org>"]
readme = "README.md"
packages = [{include = "llm_codellama", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
llm = "^0.8"
torch = "^2.0.1"
transformers = "^4.32.0"
accelerate = "^0.22.0"
llama-cpp-python = "^0.1.79"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.286"
pytest = "^7.4.0"
ipython = "^8.14.0"

[tool.poetry.plugins.llm]
codellama = "llm_codellama.codellama"
codellama_gguf = "llm_codellama.codellama_gguf"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
